Port to cpp, unfinished
This ports the initial commit from the Java version over to cpp. It's a
one to one copy for large portions of the code.

The binary search algorithm presented by the teacher has been ported as
well.

The quickSort algorithm is still work in progress. The partitioning is
not yet implemented.